Enum car_compress::Format

pub enum Format {
    LZW(Quality),
    LZH(Quality),
    Gzip(Quality),
    Zip7(Quality),
    Bzip2(Quality),
    Xz(Quality),
    Brotli(Quality),
    Lz4(Quality),
    Snappy(Quality),
    Zstd(Quality),
    Tar(Quality),
}

Describes the compression algortim we're working with.

This can be detected, or set by the user depending on if they're compressing or decompressing.

Variants
LZW(Quality)LZH(Quality)Gzip(Quality)Zip7(Quality)Bzip2(Quality)Xz(Quality)Brotli(Quality)Lz4(Quality)Snappy(Quality)Zstd(Quality)Tar(Quality)
Methods

impl Format

pub fn get_extension(&self) -> &'static str

Returns the suggested or common extension for a file format this isn't a hard and fast rule, some are identical.

pub fn from_path<P: AsRef<Path>>(p: P) -> Result<Format>

Try to find out the format of a file

This will attempt to open the file at path and read the first 16bytes matching that against a known magic number.

If the file's type is unknown this method will return Err(InvalidInput)

pub fn from_reader<R: Read + Seek>(r: &mut R) -> Result<Format>

A slightly more efficient way to find a file's type.

This method will seek to the start, read 16 bytes, then seek back to the start. The goal of this is to avoid multiple open/close cycles

If the file's type is unknown this method will return Err(InvalidInput)